Fixed compilation issues (mostly).
